# swagger.yaml
openapi: 3.0.3
info:
  title: EPOS Backoffice RESTful APIs
  description: This is the Backoffice RESTful APIs Swagger page.
  contact:
    email: apis@lists.epos-ip.org
  license:
    name: MIT License
    url: https://epos-ci.brgm.fr/epos/WebApi/raw/master/LICENSE
  version: 2.0.0
servers:
  - url: http://localhost:8080/api/backoffice/v1
    description: Inferred Url
paths:
...
  /distribution:
    post:
      tags:
        - V3 - Distribution
      summary: "Create Distribution instance"
      description: ""
      operationId: "distributionpost"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/distribution'
      responses:
        '201':
          description: Created.
        '204':
          description: No content.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    put:
      tags:
        - V3 - Distribution
      summary: "Update Distribution"
      description: ""
      operationId: "distributionputstate"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/distribution'
      responses:
        '200':
          description: ok.
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
  /distribution/{instance_id}:
    get:
      tags:
        - V3 - Distribution
      summary: "Retrive Distribution"
      description: ""
      operationId: "distributionget"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Instance Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/distribution'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    delete:
      tags:
        - V3 - Distribution
      summary: "Delete Distribution"
      description: ""
      operationId: "distributiondelete"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/distribution'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
...
  /operation:
    post:
      tags:
        - V3 - Operation
      summary: "Create Operation instance"
      description: ""
      operationId: "operationpost"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/operation'
      responses:
        '201':
          description: Created.
        '204':
          description: No content.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    put:
      tags:
        - V3 - Operation
      summary: "Update Operation"
      description: ""
      operationId: "operationputstate"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/operation'
      responses:
        '200':
          description: ok.
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
  /operation/{instance_id}:
    get:
      tags:
        - V3 - Operation
      summary: "Retrive Operation"
      description: ""
      operationId: "operationget"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Instance Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/operation'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    delete:
      tags:
        - V3 - Operation
      summary: "Delete Operation"
      description: ""
      operationId: "operationdelete"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "Id of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/operation'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
...
components:
  schemas:
...
    operation:
      required:
        - UID
        - template
      properties:
        UID:
          type: string
          example: uid
          description: 'This property contains the main identifier for the Operation'
        method:
          type: string
          example: get
          description: 'The HTTP method.'
        returns:
          type: array
          items:
            type: string
            example: 'json'
          description: 'This property is used to specify the output format of the Operation. The possible values are listed here: https://www.iana.org/assignments/media-types/media-types.xhtml'
        template:
          type: string
          example: 'uri template'
          description: 'The syntax of the template literal is specified by the [RFC6570] URI Template syntax.'
        spatialExtent:
          type: array
          items:
            $ref: '#/components/schemas/mapping'
          description: 'A variable-to-property mapping of the IRI template.'
    distribution:
      required:
        - UID
        - accessURL
      type: object
      properties:
        UID:
          type: string
          example: uid
          description: 'This property contains the main identifier for the Distribution.'
        accessService:
          type: string
          example: type
          description: 'This property refers to the WebService which supports selection of an extract, sub-set, or combination of data'
        accessURL:
          type: array
          items:
            type: string
            example: 'url'
          description: 'This property contains a URL that gives access to a Distribution of the Data Product. The resource at the access URL may contain information about how to get the Data Product.'
        description:
          type: array
          items:
            type: string
            example: 'description'
          description: 'This property contains a free-text account of the Distribution. This property can be repeated for parallel language versions of the description.'
        downloadURL:
          type: array
          items:
            type: string
            example: 'url'
          description: 'This property contains a URL that is a direct link to a downloadable file in a given format.'
        format:
          type: string
          example: format
          description: 'This property refers to the type of the Distribution. A controlled vocabulary for the values has not been established.'
        issued:
          type: string
          format: date-time
          description: 'This property contains the date of formal issuance (e.g., publication) of the Distribution.'
        licence:
          type: string
          example: licence
          description: 'This property refers to the licence under which the Distribution is made available.'
        modified:
          type: string
          format: date-time
          description: 'This property contains the most recent date on which the Distribution was changed or modified.'
        title:
          type: array
          items:
            type: string
            example: 'title'
          description: 'This property contains a name given to the Distribution. This property can be repeated for parallel language versions of the description.'
        type:
          type: string
          example: type
          description: 'This property refers to the type of the Distribution.'
    mapping:
      required:
        - variable
        - label
        - required
        - range
      properties:
        defaultValue:
          type: string
          example: 'default Value'
          description: 'This property contains the default value of the parameter.'
        label:
          type: string
          example: 'label'
          description: 'This property contains a short string used to describe the meaning of the parameter.'
        maxValue:
          type: string
          example: 'max Value'
          description: 'This property contains the maximum value of the parameter.'
        minValue:
          type: string
          example: 'min Value'
          description: 'This property contains the minimum value of the parameter.'
        paramValue:
          type: array
          items:
            type: string
            example: 'Param Value'
          description: 'This property contains one of the possible values which should be used in the web service query. It could be repeated as many times as needed.'
        property:
          type: string
          example: 'property'
          description: 'This property contains the vocabulary term which indicates the semantic description of parameter.'
        range:
          type: string
          example: 'range'
          description: 'This property contains the type of parameter'
        required:
          type: string
          example: 'required'
          description: 'This property contains true if the property is required, false otherwise.'
        valuePattern:
          type: string
          example: 'value Pattern'
          description: 'This property contains the regular expression for testing values according to the parameters specification.'
        variable:
          type: string
          example: 'variable'
          description: 'This property contains the name of the parameter as required by web service specifications.'
...
  securitySchemes:
    aai:
      type: apiKey
      in: header
      name: Authorization